Implement the CLI interface
