+++ cmd/examplar/.gitignore
@@ -1,2 +1,3 @@
 examplar
 trace_test.go
+integration
